Fixed Retrofit issue

+++ app/src/main/java/com/example/girish/bake/Model/Ingredient_Result/Ingredient.java
@@ -10,7 +10,7 @@ import com.google.gson.annotations.SerializedName;
 public class Ingredient {
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private float quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -18,11 +18,11 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public Integer getQuantity() {
+    public float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
